chore: add minimal CI; fix: complete BFS implementation (#3)
**Note:** This MR was largely generated by Claude and has not been
completely reviewed by me (the human). You should feel free to defer
your review until this warning has been removed.

## Summary
rosalind-rust is a personal exercise repo. This PR:

- Adds minimal scaffolding: `.github/workflows/check.yml` (build + test
only — no clippy or fmt), `.github/dependabot.yml`,
`.github/CODEOWNERS`.
- **Fixes a real compilation error**:
`search::bfs::breadth_first_search` was a half-written stub (queue +
visited set, no return value). Completed with the standard queue-based
BFS algorithm, returning `Vec<i64>` of shortest-path distances from
`start` to vertices `1..=n`, with `-1` for unreachable vertices
(matching the Rosalind BFS problem spec in the module doc).
- Widens `Answer::IntVec` from `Vec<usize>` to `Vec<i64>` so the `-1`
sentinel can be represented directly.

## Test plan
- [x] `cargo build --all-targets` passes.
- [x] `cargo test` passes (0 tests, but no compilation errors).
- [ ] *(no test added for BFS — repo has no test infrastructure; a
proper test would require encoding a Rosalind test case, which seemed
out of scope for a CI-enablement PR)*

## Commits
1. `chore: add minimal CI, dependabot, CODEOWNERS`
2. `chore: relax CI to build+test only`
3. `fix: complete BFS implementation in search::bfs`

---------

+++ src/lib.rs
@@ -44,7 +44,7 @@ pub enum Answer {
     String(String),
     Int(usize),
     StringVec(Vec<String>),
-    IntVec(Vec<usize>),
+    IntVec(Vec<i64>),
     CountAndIter {
         count: usize,
         iter: Box<dyn Iterator<Item = String>>,
